+++ src/components/Pages/Home/Home.jsx
@@ -1,0 +1,17 @@
+import React from 'react';
+
+const Home = () => {
+  return (
+    <div className="App">
+          <header className="App-header">
+            <p>Edit <code>src/App.js</code> and save to reload.</p>
+            <a className="App-link" href="https://reactjs.org" target="_blank" rel="noopener noreferrer">
+              Learn React
+            </a>
+          </header>
+        </div>
+      );
+}
+    
+export default  Home; 
+
