+++ src/components/Pages/NotFoundPage/NotFoundPage.jsx
@@ -1,0 +1,14 @@
+
+import React from 'react';
+
+const NotFoundPage = () => {
+  
+    return (
+      <div>
+        <h1>NOT FOUND</h1>
+      </div>
+    );
+  }
+    
+export default  NotFoundPage; 
+
