+++ src/components/Pages/Project/Project.jsx
@@ -1,0 +1,15 @@
+import React from 'react';
+import { useParams } from 'react-router';
+
+const Project = () => {
+    const { id } = useParams();
+  
+    return (
+      <div>
+        <h3>ID: {id}</h3>
+      </div>
+    );
+  }
+    
+export default  Project; 
+
